requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.3.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0
boto3>=1.28.0
